+++ poja/core.py
@@ -1,25 +1,27 @@
 import poja.mygit as mygit
 import shutil
 import poja.sed as sed
-from poja.myrich import print_title, print_normal, print_banner
+from poja.myrich import print_title, print_normal, print_banner, print_warn
 from poja.version import get_version
 from poja.vpcscoped import set_vpc_scoped_resources
 from poja.genclients import set_gen_clients
 from poja.database import set_postgres, set_sqlite
 import yaml
+from yaml.loader import BaseLoader
 import os
-import platform
+from poja.myos import cd_then_exec
 from pathlib import Path
 
 GIT_URL = "https://github.com/hei-school/poja"
-GIT_TAG_OR_COMMIT = "bca09e8"
+GIT_TAG_OR_COMMIT = "35630b0"
 
 DEFAULT_GROUP_NAME = "school.hei"
 DEFAULT_PACKAGE_FULL_NAME = DEFAULT_GROUP_NAME + ".poja"
 
 
 def gen(
-    app_name,
+    poja_conf=None,
+    app_name=None,
     region="eu-west-3",
     with_own_vpc="false",
     ssm_sg_id=None,
@@ -43,6 +45,47 @@ def gen(
     worker_batch=5,
     with_snapstart="false",
 ):
+    if poja_conf is not None:
+        with open(poja_conf, "r") as conf_strem:
+            conf = yaml.load(conf_strem, Loader=BaseLoader)
+            if get_version() != conf["cli_version"]:
+                raise Exception(
+                    "You must use the poja version defined in your conf file"
+                )
+            print_warn(
+                "Only --poja-conf will be taken into account: all other arguments will be ignored!"
+            )
+            print_warn(
+                "No default value will be used: explicit everything in your conf file!!"
+            )
+            app_name = conf["app_name"]
+            region = conf["region"]
+            with_own_vpc = conf["with_own_vpc"]
+            ssm_sg_id = conf["ssm_sg_id"]
+            ssm_subnet1_id = conf["ssm_subnet1_id"]
+            ssm_subnet2_id = conf["ssm_subnet2_id"]
+            ses_source = conf["ses_source"]
+            with_swagger_ui = conf["with_swagger_ui"]
+            package_full_name = conf["package_full_name"]
+            custom_java_repositories = conf["custom_java_repositories"]
+            custom_java_deps = conf["custom_java_deps"]
+            custom_java_env_vars = conf["custom_java_env_vars"]
+            with_gen_clients = conf["with_gen_clients"]
+            with_database = conf["with_database"]
+            jacoco_min_coverage = conf["jacoco_min_coverage"]
+            with_publish_to_npm_registry = conf["with_publish_to_npm_registry"]
+            ts_client_default_openapi_server_url = conf[
+                "ts_client_default_openapi_server_url"
+            ]
+            ts_client_api_url_env_var_name = conf["ts_client_api_url_env_var_name"]
+            frontal_memory = int(conf["frontal_memory"])
+            worker_memory = int(conf["worker_memory"])
+            worker_batch = int(conf["worker_batch"])
+            with_snapstart = conf["with_snapstart"]
+
+    if app_name is None:
+        raise Exception("app_name must be defined")
+
     if output_dir is None:
         output_dir = app_name
 
@@ -51,15 +94,15 @@ def gen(
     print_title("Checkout base repository...")
     print_normal("git_url=%s" % GIT_URL)
     print_normal("git_tag=%s" % GIT_TAG_OR_COMMIT)
-    temp_dir = mygit.checkout(GIT_URL, GIT_TAG_OR_COMMIT, no_git=True)
-    print_normal("temp_dir=%s" % temp_dir)
+    tmp_dir = mygit.checkout(GIT_URL, GIT_TAG_OR_COMMIT, no_git=True)
+    print_normal("tmp_dir=%s" % tmp_dir)
 
     print_title("Handle arguments...")
     exclude = "*.jar"
     print_normal("region")
-    sed.find_replace(temp_dir, "<?aws-region>", region, exclude)
+    sed.find_replace(tmp_dir, "<?aws-region>", region, exclude)
     print_normal("ses_source")
-    sed.find_replace(temp_dir, "<?aws-ses-source>", ses_source, exclude)
+    sed.find_replace(tmp_dir, "<?aws-ses-source>", ses_source, exclude)
 
     if with_snapstart == "true":
         if with_database == "sqlite":
@@ -73,91 +116,91 @@ def gen(
     else:
         function_snapstart = ""
         function_snapstart_java_env = "JAVA_TOOL_OPTIONS: -XX:+TieredCompilation -XX:TieredStopAtLevel=1 -Dspring.main.lazy-initialization=true -Dspring.data.jpa.repositories.bootstrap-mode=lazy -Dspring.datasource.max-active=5 -Dspring.datasource.max-idle=1 -Dspring.datasource.min-idle=1 -Dspring.datasource.initial-size=1"
-    sed.find_replace(temp_dir, "<?function-snapstart>", function_snapstart, exclude)
+    sed.find_replace(tmp_dir, "<?function-snapstart>", function_snapstart, exclude)
     sed.find_replace(
-        temp_dir,
+        tmp_dir,
         "<?function-snapstart-java-env-vars>",
         function_snapstart_java_env,
         exclude,
     )
 
     print_normal("frontal_memory")
-    sed.find_replace(temp_dir, "<?frontal-memory>", str(frontal_memory), exclude)
+    sed.find_replace(tmp_dir, "<?frontal-memory>", str(frontal_memory), exclude)
     print_normal("worker_memory")
-    sed.find_replace(temp_dir, "<?worker-memory>", str(worker_memory), exclude)
+    sed.find_replace(tmp_dir, "<?worker-memory>", str(worker_memory), exclude)
     print_normal("worker_batch")
-    sed.find_replace(temp_dir, "<?worker-batch>", str(worker_batch), exclude)
+    sed.find_replace(tmp_dir, "<?worker-batch>", str(worker_batch), exclude)
 
     print_normal("with_database")
-    set_postgres(with_database, temp_dir, exclude)
-    set_sqlite(with_database, package_full_name, temp_dir, exclude)
+    set_postgres(with_database, tmp_dir, exclude)
+    set_sqlite(with_database, package_full_name, tmp_dir, exclude)
 
     print_normal("with_own_vpc")
     set_vpc_scoped_resources(
-        with_own_vpc, ssm_sg_id, ssm_subnet1_id, ssm_subnet2_id, temp_dir, exclude
+        with_own_vpc, ssm_sg_id, ssm_subnet1_id, ssm_subnet2_id, tmp_dir, exclude
     )
 
     print_normal("with_gen_clients")
-    set_gen_clients(with_gen_clients, temp_dir, exclude)
+    set_gen_clients(with_gen_clients, tmp_dir, exclude)
 
     print_normal("with_swagger_ui")
     if with_swagger_ui == "true":
         springdoc_java_dep = "implementation 'org.springdoc:springdoc-openapi-ui:1.7.0'"
     else:
         springdoc_java_dep = ""
-    sed.find_replace(temp_dir, "<?java-deps-springdoc>", springdoc_java_dep, exclude)
+    sed.find_replace(tmp_dir, "<?java-deps-springdoc>", springdoc_java_dep, exclude)
 
     print_normal("package_full_name")
-    sed.find_replace(temp_dir, DEFAULT_PACKAGE_FULL_NAME, package_full_name, exclude)
+    sed.find_replace(tmp_dir, DEFAULT_PACKAGE_FULL_NAME, package_full_name, exclude)
     sed.find_replace(
-        temp_dir,
+        tmp_dir,
         DEFAULT_GROUP_NAME,
         group_name_from_package_full_name(package_full_name),
         exclude,
     )
     print_normal("custom_java_repositories")
     java_repositories = replace_with_file_content(
-        temp_dir, "<?java-repositories>", custom_java_repositories, exclude
+        tmp_dir, "<?java-repositories>", custom_java_repositories, exclude
     )
-    set_package_dirs(temp_dir, package_full_name, "main")
-    set_package_dirs(temp_dir, package_full_name, "test")
+    set_package_dirs(tmp_dir, package_full_name, "main")
+    set_package_dirs(tmp_dir, package_full_name, "test")
     print_normal("custom_java_deps")
     java_deps = replace_with_file_content(
-        temp_dir, "<?java-deps>", custom_java_deps, exclude
+        tmp_dir, "<?java-deps>", custom_java_deps, exclude
     )
     print_normal("custom_java_env_vars")
     indent = "        "
     java_env_vars = replace_with_file_content(
-        temp_dir, "<?java-env-vars>", custom_java_env_vars, exclude, joiner=indent
+        tmp_dir, "<?java-env-vars>", custom_java_env_vars, exclude, joiner=indent
     )
 
     print_normal("app_name")
-    sed.find_replace(temp_dir, "<?app-name>", app_name, exclude)
+    sed.find_replace(tmp_dir, "<?app-name>", app_name, exclude)
     print_normal("jacoco_min_coverage")
     sed.find_replace(
-        temp_dir, "<?jacoco-min-coverage>", str(jacoco_min_coverage), exclude
+        tmp_dir, "<?jacoco-min-coverage>", str(jacoco_min_coverage), exclude
     )
     if with_publish_to_npm_registry == "true":
         print_normal("ts_client_default_openapi_server_url")
         sed.find_replace(
-            temp_dir,
+            tmp_dir,
             "<?ts-client-default-openapi-server-url>",
             ts_client_default_openapi_server_url,
             exclude,
         )
         print_normal("ts_client_api_url_env_var_name")
         sed.find_replace(
-            temp_dir,
+            tmp_dir,
             "<?ts-client-api-url-env-var-name>",
             ts_client_api_url_env_var_name,
             exclude,
         )
     else:
-        os.remove("%s/.github/workflows/publish-client.yml" % temp_dir)
+        os.remove("%s/.github/workflows/publish-client.yml" % tmp_dir)
 
     print_title("Save conf...")
     save_conf(
-        temp_dir,
+        tmp_dir,
         app_name,
         region,
         with_own_vpc,
@@ -173,8 +216,10 @@ def gen(
         with_gen_clients,
         with_database,
         jacoco_min_coverage,
+        with_publish_to_npm_registry,
         ts_client_default_openapi_server_url,
         ts_client_api_url_env_var_name,
+        with_snapstart,
         frontal_memory,
         worker_memory,
         worker_batch,
@@ -183,20 +228,17 @@ def gen(
 
     print_title("Rm project-specific files...")
     print_normal("README.md")
-    os.remove(temp_dir + "/README.md")
+    os.remove(tmp_dir + "/README.md")
     print_normal("application.properties")
-    os.remove(temp_dir + "/src/main/resources/application.properties")
+    os.remove(tmp_dir + "/src/main/resources/application.properties")
     print_normal("gradle.properties")
-    os.remove(temp_dir + "/gradle.properties")
+    os.remove(tmp_dir + "/gradle.properties")
 
     print_title("Format...")
-    if "Windows" in platform.system():
-        os.system("cd /D %s && format.bat" % temp_dir)
-    else:
-        os.system("cd %s && ./format.sh" % temp_dir)
+    cd_then_exec(tmp_dir, "format.bat", "./format.sh")
 
     print_title("Copy to output dir...")
-    shutil.copytree(temp_dir, output_dir, dirs_exist_ok=True)
+    shutil.copytree(tmp_dir, output_dir, dirs_exist_ok=True)
 
     print_title("Client generation...")
     print_normal("doc/api.yml")
@@ -213,7 +255,7 @@ def group_name_from_package_full_name(package_full_name):
 
 
 def save_conf(
-    temp_dir,
+    tmp_dir,
     app_name,
     region,
     with_own_vpc,
@@ -229,8 +271,10 @@ def save_conf(
     with_gen_clients,
     with_database,
     jacoco_min_coverage,
+    with_publish_to_npm_registry,
     ts_client_default_openapi_server_url,
     ts_client_api_url_env_var_name,
+    with_snapstart,
     frontal_memory,
     worker_memory,
     worker_batch,
@@ -254,31 +298,33 @@ def save_conf(
         "custom_java_env_vars": custom_java_env_vars_filename,
         "with_gen_clients": with_gen_clients,
         "with_database": with_database,
-        "jacoco-min-coverage": jacoco_min_coverage,
+        "jacoco_min_coverage": jacoco_min_coverage,
+        "with_publish_to_npm_registry": with_publish_to_npm_registry,
         "ts_client_default_openapi_server_url": ts_client_default_openapi_server_url,
         "ts_client_api_url_env_var_name": ts_client_api_url_env_var_name,
+        "with_snapstart": with_snapstart,
         "frontal_memory": frontal_memory,
         "worker_memory": worker_memory,
         "worker_batch": worker_batch,
     }
-    with open(temp_dir + "/poja.yml", "w") as conf_file:
+    with open(tmp_dir + "/poja.yml", "w") as conf_file:
         yaml.dump(conf, conf_file)
 
     print_normal(custom_java_repositories_filename)
     with open(
-        "%s/%s" % (temp_dir, custom_java_repositories_filename), "w"
+        "%s/%s" % (tmp_dir, custom_java_repositories_filename), "w"
     ) as custom_java_repositories_file:
         custom_java_repositories_file.write(custom_java_repositories)
 
     print_normal(custom_java_deps_filename)
     with open(
-        "%s/%s" % (temp_dir, custom_java_deps_filename), "w"
+        "%s/%s" % (tmp_dir, custom_java_deps_filename), "w"
     ) as custom_java_deps_file:
         custom_java_deps_file.write(custom_java_deps)
 
     print_normal(custom_java_env_vars_filename)
     with open(
-        "%s/%s" % (temp_dir, custom_java_env_vars_filename), "w"
+        "%s/%s" % (tmp_dir, custom_java_env_vars_filename), "w"
     ) as custom_java_env_vars_file:
         custom_java_env_vars_file.write(
             "\n".join([s.strip() for s in custom_java_env_vars.split("\n")])
@@ -297,28 +343,28 @@ def replace_with_file_content(
     return content
 
 
-def set_package_dirs(temp_dir, package_full_name, scope):
+def set_package_dirs(tmp_dir, package_full_name, scope):
     package_full_name_parts = get_package_full_name_parts(package_full_name)
     default_package_full_name_parts = DEFAULT_PACKAGE_FULL_NAME.split(".")
     os.rename(
-        "%s/src/%s/java/%s" % (temp_dir, scope, default_package_full_name_parts[0]),
-        "%s/src/%s/java/%s" % (temp_dir, scope, package_full_name_parts[0]),
+        "%s/src/%s/java/%s" % (tmp_dir, scope, default_package_full_name_parts[0]),
+        "%s/src/%s/java/%s" % (tmp_dir, scope, package_full_name_parts[0]),
     )
     os.rename(
         "%s/src/%s/java/%s/%s"
         % (
-            temp_dir,
+            tmp_dir,
             scope,
             package_full_name_parts[0],
             default_package_full_name_parts[1],
         ),
         "%s/src/%s/java/%s/%s"
-        % (temp_dir, scope, package_full_name_parts[0], package_full_name_parts[1]),
+        % (tmp_dir, scope, package_full_name_parts[0], package_full_name_parts[1]),
     )
     os.rename(
         "%s/src/%s/java/%s/%s/%s"
         % (
-            temp_dir,
+            tmp_dir,
             scope,
             package_full_name_parts[0],
             package_full_name_parts[1],
@@ -326,7 +372,7 @@ def set_package_dirs(temp_dir, package_full_name, scope):
         ),
         "%s/src/%s/java/%s/%s/%s"
         % (
-            temp_dir,
+            tmp_dir,
             scope,
             package_full_name_parts[0],
             package_full_name_parts[1],