+++ poja/myrich/__init__.py
@@ -1,7 +1,7 @@
 from rich.console import Console
 from rich.theme import Theme
 
-theme = Theme({"title": "bold green", "banner": "reverse"})
+theme = Theme({"title": "bold green", "banner": "reverse", "warn": "bold red"})
 console = Console(theme=theme)
 
 
@@ -11,6 +11,10 @@ def print_banner(str):
 
 def print_title(str):
     console.print("\n" + str, style="title")
+
+
+def print_warn(str):
+    console.print("\n" + str, style="warn")
 
 
 def print_normal(str):