+++ poja/vpcscoped/__init__.py
@@ -3,11 +3,11 @@ from poja.myrich import print_normal
 
 
 def set_vpc_scoped_resources(
-    with_own_vpc, ssm_sg_id, ssm_subnet1_id, ssm_subnet2_id, temp_dir, exclude
+    with_own_vpc, ssm_sg_id, ssm_subnet1_id, ssm_subnet2_id, tmp_dir, exclude
 ):
     if with_own_vpc == "true":
         sed.find_replace(
-            temp_dir,
+            tmp_dir,
             "<?function-vpc-config>",
             """VpcConfig:
       SecurityGroupIds:
@@ -18,7 +18,7 @@ def set_vpc_scoped_resources(
             exclude,
         )
         sed.find_replace(
-            temp_dir,
+            tmp_dir,
             "<?db-subnet-group>",
             """DBSubnetGroup:
     Type: AWS::RDS::DBSubnetGroup
@@ -31,13 +31,13 @@ def set_vpc_scoped_resources(
             exclude,
         )
         sed.find_replace(
-            temp_dir,
+            tmp_dir,
             "<?db-subnet-group-name>",
             "DBSubnetGroupName: !Ref DBSubnetGroup",
             exclude,
         )
         sed.find_replace(
-            temp_dir,
+            tmp_dir,
             "<?db-sg-ids>",
             """VpcSecurityGroupIds:
         - !Sub '{{resolve:ssm:<?ssm-param-sg-id>}}'""",
@@ -45,19 +45,19 @@ def set_vpc_scoped_resources(
         )
 
         print_normal("ssm_sg_id")
-        sed.find_replace(temp_dir, "<?ssm-param-sg-id>", ssm_sg_id, exclude)
+        sed.find_replace(tmp_dir, "<?ssm-param-sg-id>", ssm_sg_id, exclude)
         print_normal("ssm_subnet1_id")
         sed.find_replace(
-            temp_dir, "<?ssm-param-name-subnet1-id>", ssm_subnet1_id, exclude
+            tmp_dir, "<?ssm-param-name-subnet1-id>", ssm_subnet1_id, exclude
         )
         print_normal("ssm_subnet2_id")
         sed.find_replace(
-            temp_dir, "<?ssm-param-name-subnet2-id>", ssm_subnet2_id, exclude
+            tmp_dir, "<?ssm-param-name-subnet2-id>", ssm_subnet2_id, exclude
         )
     else:
-        sed.find_replace(temp_dir, "<?function-vpc-config>", "", exclude)
-        sed.find_replace(temp_dir, "<?db-sg-ids>", "", exclude)
+        sed.find_replace(tmp_dir, "<?function-vpc-config>", "", exclude)
+        sed.find_replace(tmp_dir, "<?db-sg-ids>", "", exclude)
 
         # TODO: Aurora Postgres v1 cannot be put inside AWS-managed network
-        sed.find_replace(temp_dir, "<?db-subnet-group>", "", exclude)
-        sed.find_replace(temp_dir, "<?db-subnet-group-name>", "", exclude)
+        sed.find_replace(tmp_dir, "<?db-subnet-group>", "", exclude)
+        sed.find_replace(tmp_dir, "<?db-subnet-group-name>", "", exclude)