+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+# [10.1.0](https://github.com/hei-school/poja-cli/compare/v10.0.0...v10.1.0) (2024-01-04)
+
+
+### Features
+
+* --poja-conf ([8c9fb98](https://github.com/hei-school/poja-cli/commit/8c9fb98b0cf4e7117d68bf9125e275c6634389f7))
+
+
+
+# [10.0.0](https://github.com/hei-school/poja-cli/compare/v9.1.1...v10.0.0) (2024-01-03)
+
+
+### Bug Fixes
+
+* --with-database=none --> non-poja-managed-postgres ([1bdfcd1](https://github.com/hei-school/poja-cli/commit/1bdfcd149b6c97ef5bf0f7a1fa404d17394494a5))
+
+
+### BREAKING CHANGES
+
+* --with-database=none --> non-poja-managed-postgres
+
+
+
+## [9.1.1](https://github.com/hei-school/poja-cli/compare/v9.1.0...v9.1.1) (2024-01-03)
+
+
+### Bug Fixes
+
+* specify region for S3AsyncClient bean ([81386c2](https://github.com/hei-school/poja-cli/commit/81386c26edbb9c2820f335fed389fe480c9ad799))
+
+
+
+# [9.1.0](https://github.com/hei-school/poja-cli/compare/v9.0.3...v9.1.0) (2024-01-03)
+
+
+### Features
+
+* specify SesClient region ([86a211b](https://github.com/hei-school/poja-cli/commit/86a211b54ac33eeb7fdb4d05f311a3b2dee57056))
+
+
+
 ## [9.0.3](https://github.com/hei-school/poja-cli/compare/v9.0.2...v9.0.3) (2023-12-29)